refactor(tools): separate Team Apply assessment from presentation
